Corrección de tiempos de caducidad y actualización de logs
- Se actualizan los tiempos de caducidad para que se mantengan en segundos en el fichero de configuración y se procesen como milisegundos en todo el sistema.
- Se corrigen errores en los tiempos de caducidad derivados de la mezcla de segundos y milisegundos.
- Se agregan nuevas trazas de log en los servicios del componente central.
- Se modifican los logs de la aplicación para marcar muchos de ellos como de debug.

+++ fire-signature/src/main/java/es/gob/fire/server/services/internal/ExpiredSessionCleanerThread.java
@@ -19,17 +19,17 @@ public class ExpiredSessionCleanerThread extends Thread {
 
 	private final String[] ids;
 	private final Map<String, FireSession> sessions;
-	private final int timeout;
+	private final long timeout;
 
 	/**
 	 * Construye el objeto para la eliminaci&oacute;n de sesiones caducadas.
 	 * @param ids Identificadores de las sesiones que se tienen que evaluar.
 	 * @param sessions Mapa con todas las sesiones.
-	 * @param tempTimeout Tiempo de caducidad de los ficheros temporales.
+	 * @param tempTimeout Tiempo de caducidad en milisegundos de los ficheros temporales.
 	 */
 	public ExpiredSessionCleanerThread(final String[] ids,
 			final Map<String, FireSession> sessions,
-			final int tempTimeout) {
+			final long tempTimeout) {
 		this.ids = ids;
 		this.sessions = sessions;
 		this.timeout = tempTimeout;
